Added comments and edited code
I changed the "val  jumpCount = mutableMapOf<UUID, Int>()" in the class NoJumping allowed on line 13  to the same thing but with private at the beginning. I also changed the code of the class NoSprintingAllowed from canceling the event, which didn't work, to setting the players hunger bars to 3. I also added comments to all the code.